More float/double problems, especially in the browser viewer, which was
2007-10-01  Bastien Nocera  <hadess@hadess.net>

	* browser-plugin/totem-plugin-viewer.c: (on_tick):
	* src/backend/bacon-video-widget-xine.c:
	(bacon_video_widget_tick_send): More float/double problems,
	especially in the browser viewer, which was causing the seekbar
	to yo-yo with some streams


svn path=/trunk/; revision=4738

+++ browser-plugin/totem-plugin-viewer.c
@@ -1557,7 +1557,7 @@ static void
 on_tick (GtkWidget *bvw,
 		gint64 current_time,
 		gint64 stream_length,
-		float current_position,
+		double current_position,
 		gboolean seekable,
 		TotemEmbedded *emb)
 {
